+++ src/main/java/net/raphimc/viabedrock/protocol/types/ByteArrayType.java
@@ -37,7 +37,7 @@ public class ByteArrayType extends Type<byte[]> {
 
     @Override
     public byte[] read(ByteBuf buffer) {
-        final int length = this.length == -1 ? (int) BedrockTypes.UNSIGNED_VAR_INT.readPrimitive(buffer) : this.length;
+        final int length = this.length == -1 ? BedrockTypes.UNSIGNED_VAR_INT.readPrimitive(buffer) : this.length;
         Preconditions.checkArgument(buffer.isReadable(length), "Length is fewer than readable bytes");
         final byte[] array = new byte[length];
         buffer.readBytes(array);
